Code inserted into functions. Still need to specify the types for mypy
